Problem: We need to enable LSEED to run as a Kubernetes pod/service
LSEED uses port 53 (DNS) to provide joining LND nodes with a list of
existing nodes on the Lightning network. Unfortunately services on port
53 are a bit difficult to provide behind the regular Kubernetes LB or
ingress controller. In particular the fact that LSEED answers queries
on both the UDP and TCP protocol makes it difficult to run in a
Kubernetes environment.

To solve the issues describe above this commit provides the following
code changes:

- enhanced logging for events that add new nodes to the LSEED node list
- a Dockerfile to create a containerized version of LSEED
- new parameters with separate listener addresses for UDP and TCP
- a separate listener for UDP and TCP requests utilizing these new
  parameters
- a couple of typo fixes to the Debugf statements in lseed and
  tests